fix(ci): first-party line limit — split two oversized unit test modules over 1500 lines
